refactor(mapgen): collapse observational input ownership (#2107)
refactor(mapgen): collapse observational input ownership

refactor(mapgen): collapse executable operation bindings

refactor(mapgen): collapse artifact capability projection

refactor(mapgen): admit artifacts once

refactor(mapgen): unify step dependencies

refactor(mapgen): reduce effects to plan completions

refactor(mapgen): move initial setup into step context

refactor(mapgen): name step observations honestly

refactor(mapgen): split precipitation refinement authority

refactor(mapgen): preserve canonical step identity

refactor(mapgen): unify artifact dependency analysis

+++ mods/mod-swooper-maps/src/domain/morphology/modules/coasts/ops/compute-coastal-adjacency/strategies/wrapped-hex-adjacency/index.ts
@@ -9,7 +9,7 @@ export default createStrategy(ComputeCoastalAdjacencyContract, StrategyDefinitio
   run: (input) => {
     const { width, height } = input;
     const size = width * height;
-    const landMask = input.landMask as Uint8Array;
+    const landMask = input.landMask;
 
     const coastalLand = new Uint8Array(size);
     const coastalWater = new Uint8Array(size);
